requirements.txt
Flask
flask-limiter
requests
python-dotenv